% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchEfficiency.R
\name{accuracy_F}
\alias{accuracy_F}
\title{Calculate Accuracy}
\usage{
accuracy_F(coverage, capacityPlan)
}
\arguments{
\item{coverage}{\itemize{
\item the per interval number of agents scheduled
}}

\item{capacityPlan}{\itemize{
\item the per-interval headcount requirement
}}
}
\description{
This function loads two dataframes and calculates the percentage simmilarity between their
corresponding elements. It is important that the dataframes are of the same size and column ids
}
\details{
Inputs: (coverage:dataframe,capacityPlan:dataframe)
}
