% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sch2Coverage.R
\name{coverage}
\alias{coverage}
\title{Calculate Coverage}
\usage{
coverage(schedule, nDays = 7, intervals = 24)
}
\arguments{
\item{schedule}{\itemize{
\item a dataframe with fields startDay, endDay, startTime and endTime all in integers
}}

\item{nDays}{\itemize{
\item the number of days considered
}}

\item{intervals}{\itemize{
\item the number of intervals within the day (24 if hourly data)
}}
}
\description{
This function loads a dataframe with schedule and calculates the per - interval per day coverage
}
\details{
Inputs: (schedule:dataframe,nDays:int,intervals:int)
}
