% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{erlang_c}
\alias{erlang_c}
\title{Calculate Erlangs using Erlang C formula}
\usage{
erlang_c(agents, volume, AHT, interval = 1)
}
\arguments{
\item{agents}{\itemize{
\item the number agents available in that interval
}}

\item{volume}{\itemize{
\item the number of expected contacts in the interval
}}

\item{AHT}{\itemize{
\item the average handle time per contact
}}

\item{interval}{\itemize{
\item the number of time units  where the volume is expected
}}
}
\description{
This function calculates the number of Erlangs in a given interval
}
\details{
Inputs: (agents: int, volume:int,AHT:float,interval:int)
}
