% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{erlang_cl}
\alias{erlang_cl}
\title{Calculate Erlang_CL}
\usage{
erlang_cl(agents, volume, AHT, interval = 1, lines)
}
\arguments{
\item{agents}{\itemize{
\item the number agents available in that interval
}}

\item{volume}{\itemize{
\item the number of expected contacts in the interval
}}

\item{AHT}{\itemize{
\item the average handle time per contact
}}

\item{interval}{\itemize{
\item the number of time units  where the volume is expected
}}

\item{lines}{\itemize{
\item total number of customers that can be in the system at the same time
}}
}
\description{
This function calculates Erlangs using Erlang C while considering the available number of lines formula
}
\details{
Inputs: (agents: int, volume:int,AHT:float,interval:int, lines: int)
}
