% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{erlang_xd}
\alias{erlang_xd}
\title{Calculate Erlang_XD}
\usage{
erlang_xd(
  agents,
  volume,
  AHT,
  interval = 1,
  lines,
  patience,
  retrials,
  definition = 2
)
}
\arguments{
\item{agents}{\itemize{
\item the number agents available in that interval
}}

\item{volume}{\itemize{
\item the number of expected contacts in the interval
}}

\item{AHT}{\itemize{
\item the average handle time per contact
}}

\item{interval}{\itemize{
\item the number of time units  where the volume is expected
}}

\item{lines}{\itemize{
\item total number of customers that can be in the system at the same time
}}

\item{patience}{\itemize{
\item The average time a customer is willing to wait in the queue
}}

\item{retrials}{\itemize{
\item The probability that a customer who abandons, redials. (0 ≤ Retrials ≤ 1)
}}

\item{definition}{\itemize{
\item modes of calculating the service level (0-virtual SLA, 1 - answered SLA, 2 - offered sla)
}}
}
\description{
This function calculates Erlangs using Erlang X considering the definition
}
\details{
Inputs: (agents: int, volume:int,AHT:float,interval:int, lines: int, patience:int, retrials: int, definition: boolean)
}
