% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{intensity}
\alias{intensity}
\title{Calculate Intensity}
\usage{
intensity(volume, AHT, interval = 1)
}
\arguments{
\item{volume}{\itemize{
\item the number of expected contacts in the interval
}}

\item{AHT}{\itemize{
\item the average handle time per contact
}}

\item{interval}{\itemize{
\item the number of time units  where the volume is expected
}}
}
\description{
This function calculates  intensity of the incoming volume (workload) in a given interval
}
\details{
Inputs: (volume:int,AHT:float,interval:int)
}
