% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{Occupancy}
\alias{Occupancy}
\title{Calculate Occupancy}
\usage{
Occupancy(volume, AHT, interval, agents)
}
\arguments{
\item{volume}{\itemize{
\item the number of expected contacts in the interval
}}

\item{AHT}{\itemize{
\item the average handle time per contact
}}

\item{interval}{\itemize{
\item the number of time units  where the volume is expected
}}

\item{agents}{\itemize{
\item the number agents available in that interval
}}
}
\description{
This function calculates the expected level of occupancy
}
\details{
Inputs: (volume:int,AHT:float,interval:int,agents: int)
}
