% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{SL}
\alias{SL}
\title{Calculate Service Level}
\usage{
SL(
  agents,
  volume,
  AHT,
  slaTarget,
  interval = 1,
  shrink = 0,
  lines = NULL,
  patience = NULL,
  retrials = NULL,
  variance = NULL,
  definition = NULL
)
}
\arguments{
\item{agents}{\itemize{
\item the number agents available in that interval
}}

\item{volume}{\itemize{
\item the number of expected contacts in the interval
}}

\item{AHT}{\itemize{
\item the average handle time per contact
}}

\item{interval}{\itemize{
\item the number of time units  where the volume is expected
}}

\item{lines}{\itemize{
\item total number of customers that can be in the system at the same time
}}

\item{patience}{\itemize{
\item The average time a customer is willing to wait in the queue
}}

\item{retrials}{\itemize{
\item The probability that a customer who abandons, redials. (0 ≤ Retrials ≤ 1)
}}

\item{variance}{\itemize{
\item the variance of the forecasted volume. To be used if the distribution is not poisson.
}}

\item{definition}{\itemize{
\item modes of calculating the service level (0-virtual SLA, 1 - answered SLA, 2 - offered sla)
}}
}
\description{
This function calculates the expected service level as percentage.
}
\details{
Inputs: (agents: int, volume:int,AHT:float,interval:int, lines: int, patience:int, retrials: int, variance: float, definition:boolean)
}
